requirements.txt
# Options V4 System Requirements

# Core dependencies (required)
pandas>=1.5.0
numpy>=1.21.0

# Optional dependencies (system will work without these)
# Uncomment to install for full functionality
# pyyaml>=6.0
python-dotenv>=0.19.0
# scipy>=1.9.0

# Database connectivity (required for production)
supabase>=2.0.0

# Real-time WebSocket dependencies (for live trading)
websockets>=11.0.2
redis>=4.5.0
asyncio-mqtt>=0.13.0

# Performance enhancements for async operations
uvloop>=0.17.0

# Note: The system is designed to work without optional dependencies
# Missing dependencies will trigger graceful fallbacks